rename general.py to token_swapper.py
